Allow force-deleting sites using a special token

+++ src/handlers/deleteSite.js
@@ -17,18 +17,21 @@ module.exports = async (request, h) => {
         const token_file = path.join(live_deploy_dir, '.dattel-delete-token');
 
         // To avoid unintentional deletes, create a file with a delete token in the webroot.
-        if (!delete_token) {
-            if (!fs.existsSync(token_file)) fs.writeFileSync(token_file, nanoid());
-            return r(
-                h,
-                `To avoid unintentional deletes, a delete token is required. Please retrieve it from /.dattel-delete-token and retry.`,
-                403
-            );
-        }
+        // This can be overriden by specifying a token of 'FORCE_UNLEASH_RHYNCHOPHORUS_FERRUGINEUS', though.
+        if (delete_token !== 'FORCE_UNLEASH_RHYNCHOPHORUS_FERRUGINEUS') {
+            if (!delete_token) {
+                if (!fs.existsSync(token_file)) fs.writeFileSync(token_file, nanoid());
+                return r(
+                    h,
+                    `To avoid unintentional deletes, a delete token is required. Please retrieve it from /.dattel-delete-token and retry.`,
+                    403
+                );
+            }
 
-        // Validate the delete token.
-        const correct_delete_token = fs.existsSync(token_file) && fs.readFileSync(token_file).toString();
-        if (delete_token !== correct_delete_token) return r(h, 'Invalid delete token.', 403);
+            // Validate the delete token.
+            const correct_delete_token = fs.existsSync(token_file) && fs.readFileSync(token_file).toString();
+            if (delete_token !== correct_delete_token) return r(h, 'Invalid delete token.', 403);
+        }
 
         // Actually do the deletion.
         await caddy.DELETE(`/id/route-${site_id}`);
